chore: TypeScript POC
