+++ package.json
@@ -61,7 +61,7 @@
   "devDependencies": {
     "@edx/browserslist-config": "1.1.1",
     "@faker-js/faker": "8.4.1",
-    "@openedx/frontend-build": "13.0.28",
+    "@openedx/frontend-build": "14.0.10",
     "@tanstack/eslint-plugin-query": "4.29.9",
     "@testing-library/jest-dom": "5.11.9",
     "@testing-library/react": "12.1.5",
